Extract shared utilities from mcp.tools
Third of the planned PRs decomposing backend/mcp/tools.py. This pass
extracts the shared business utilities used by handlers — pagination
normalization, JSON conversion, taxonomy scoring + matching, snippet
extraction, and the payload builders for naics/counsel/agreements
summary/agreement trends.

New file:
- tools/shared.py (~785 LOC) — 27 utilities incl. _normalized_page,
  _json_compatible_value, _count_metadata_payload, _interpretation_payload,
  _agreement_filter_interpretation, _TaxonomyEntry dataclass, taxonomy
  scoring (_taxonomy_entries, _score_taxonomy_entry, etc.), snippet
  extraction (_extract_text_from_xml, _focused_snippet, etc.),
  payload builders (_build_taxonomy_tree, _naics_payload, _counsel_payload,
  _agreements_summary_payload, _agreement_trends_payload).

tools/__init__.py: 3539 -> 2772 LOC.

Per-tool handlers (~1700 LOC) extract next as MCP-D. After that the
package settles around ~1000 LOC in __init__.py.
